requirements.txt
aiofiles==23.1.0
aiogram==3.0.0b7
aiogram3_calendar==0.1.2b1
aiohappyeyeballs==2.4.3
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
fastapi==0.115.4
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.0
pydantic==1.10.19
pydantic_core==2.23.4
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.2.0
sniffio==1.3.1
starlette==0.41.2
typing_extensions==4.12.2
uvicorn==0.32.0
yarl==1.17.1
